arreglando bugs con testing
arreglando bugs con testing
