+++ S303-EscapeRoom/src/test/java/EscapeRoomTesting.java
@@ -12,6 +12,7 @@ import java.sql.Timestamp;
 public class EscapeRoomTesting {
 
     @Test
+    ///Arreglado el Null de Timestamp
     public void testEscapeRoomTesting() {
         EscapeRoom escapeRoom1 = new EscapeRoomBuilder()
                 .setName("Pedos House")
